olive_app/requirements.txt
streamlit>=1.32.0
scikit-learn>=1.4.0
numpy>=1.26.0
Pillow>=10.0.0
joblib>=1.3.0
gdown>=5.1.0
